+++ components/kyma-environment-broker/internal/storage/postsql/schema.go
@@ -1,4 +1,4 @@
-package schema
+package postsql
 
 import "fmt"
