+++ components/kyma-environment-broker/internal/storage/storage.go
@@ -2,7 +2,8 @@ package storage
 
 import (
 	"github.com/kyma-incubator/compass/components/kyma-environment-broker/internal/storage/dbsession"
-	"github.com/kyma-incubator/compass/components/kyma-environment-broker/internal/storage/service"
+	"github.com/kyma-incubator/compass/components/kyma-environment-broker/internal/storage/entity"
+	"github.com/kyma-incubator/compass/components/kyma-environment-broker/internal/storage/postsql"
 	"github.com/pkg/errors"
 )
 
@@ -11,14 +12,14 @@ type BrokerStorage interface {
 }
 
 func New(connectionURL string) (BrokerStorage, error) {
-	connection, err := InitializeDatabase(connectionURL)
+	connection, err := postsql.InitializeDatabase(connectionURL)
 	if err != nil {
 		return nil, errors.Wrap(err, "Failed to initialize database")
 	}
 	fact := dbsession.NewFactory(connection)
 
 	return storage{
-		instance: service.NewInstance(fact),
+		instance: entity.NewInstance(fact),
 	}, nil
 }
